Do not emit import-syntax forms when no syntax forms are exported
This avoids importing and re-importing the same basic modules over and
over again in an eval expression (as typically, one tends to import
scheme and chicken.base, which provide quite large environments, which
exacerbates the problem).

This also exposed a minor problem with the types-db-consistency test:
It didn't import all required modules, but they were available at the
toplevel due to imports done implicitly in import libraries of those
modules it _did_ import.

This should mitigate #1457 somewhat.  Of course, if the user imports
several libraries that have lots of imports and at least one syntax
export, it will still be slow (as this doesn't fix the underlying
quadratic behaviour), but it should be nowhere near as bad as it was
before this patch.

As an added bonus, this should also increase compile times for any
code, because those evaluated imports will always run inside the
compiler when loading the import libraries, regardless of whether the
output program is static or dynamic.  For compiling CHICKEN itself
this makes little difference (due to -explicit-use), but for a simple
program that only does (import parley), compilation time dropped from
1.5s to 0.8s (with a DEBUGBUILD).

Startup time of csi also improves dramatically: csi -e '' takes 0.024s
after this patch versus 0.30s before (both with a DEBUGBUILD).

Signed-off-by: felix <felix@call-with-current-continuation.org>