Mark all extensions as static when using (linkage static)
Rather than trying to locate static extensions during compilation with
`chicken.load#find-static-extension', this leaves that task for csc to
do at the linking step, since csc.scm already has more complete logic
for locating files in `find-object-file', including a consideration of
files in the current directory and whether setup-mode is enabled. Doing
things this way ensures that objects from the build directory are found
even before they're installed into the repository, which is necessary
when compiling an egg that has dependencies from one static component to
another.

This finishes off #1468, for which a9fcaf7f was an incomplete fix.

Signed-off-by: felix <felix@call-with-current-continuation.org>