Splice includes into body contexts so definitions are handled correctly
Now that non-toplevel definitions outside a "body" context are no longer
allowed, we have to expand include forms during body canonicalisation so
that any definitions in the included file are correctly spliced into the
surrounding context. Otherwise, they won't be recognised as internal
definitions and the compiler will reject them as "toplevel definitions
in non-toplevel context".

So, whenever a `##core#include' node is encountered, it's now extended
to include the remainder of the forms in the surrounding body and
control is handed back to the compiler. Then, whenever the compiler
reads forms from an included file, it checks for a body and, if one is
present, it knows it should return to the canonicalisation routine with
the included forms (as well as the remainder of the original body
context). If no body is present, included forms are treated as usual,
i.e. as a normal sequence that gets inserted into a `##core#begin' node.
This treatment is similar to what we currently do for modules, which
must also be handled as a special case during body canonicalisation.

Signed-off-by: felix <felix@call-with-current-continuation.org>