Some small debugger data and wire protocol improvements
These changes are intended to simplify the use of the debugging protocol
by client applications by fixing a few oddities that would otherwise
need to be coded around on the client side.

Populate the "location" slot for call events. Previously, the debugging
stub would send Scheme filenames and line number information to clients
in the "location" slot for all except for 'call' events, which would
instead have the location as a prefix of the "value" slot. Move this
source information into the "location" slot so that all events sent to
the client use the fields in the same way in all cases.

Send missing values to the client as `#f' rather than as strings.
Previously, the debugging stub would send missing values to the client
as either an empty string or a string containing "#f" (a byproduct of
using `->string' during code generation), but it's easier to handle the
"real" #f token on the client side. So, introduce a `send_string_value'
procedure that sends C strings to the client as either a quoted string
or #f if the string is NULL or empty, rather than as strings in all
cases. Update call sites to indicate missing events and file locations
as NULL in C and #f on the wire. This requires bumping the predefined
integer value definitions by one, since event locations may now be NULL
if no Scheme or C source information is available. Rename `send_value'
to `send_scheme_value' for consistency with `send_string_value'. Update
feathers.tcl as necessary.

Emit C source info as a single string, rather than as a separate
filename and line number, to simplify its use in dbg-stub.c.

Use symbols for `##core#debug-event' node event types in core.scm.
Previously, these were strings in some places and symbols in others.

Signed-off-by: Peter Bex <peter@more-magic.net>