Fix remote version checking in chicken-install by resolving egg server names
Previously, the `check-remote-version' procedure would never expand egg
server aliases to their FQDNs, so its calls to `try-list-versions' would
always fail (because they'd be hitting, for example, "call-cc" and not
"code.call-cc.org"). Add location resolution to this procedure.

Also, pull the call to `resolve-location' in the `fetch-egg-sources'
procedure up into its loop initialisation, for consistency with the
other places in chicken-install.scm where we walk the server list.

Signed-off-by: Peter Bex <peter@more-magic.net>