Allow "chicken-status -cached" to be used with egg name arguments
As with the previous commit, this allows `chicken-status -cached NAME`
to list just the eggs in the cache of the given name.

Avoid blowing up when the cache contains a directory with no egg file in
it. Previously, `read-info' would signal an error, but we now return #f
and handle the missing info in the caller.

Fix a bug that would prevent versions of cached eggs from being
detected, since the `dir' passed to `list-egg-info' was one level too
deep (it included the egg name, but that is already handled by the
procedure).

Drop "(none)" from the output when no eggs are matched, so that all the
chicken-status flags that cause things to be listed work the same way.

Signed-off-by: Peter Bex <peter@more-magic.net>