Two small fixes for egg refetching behaviour
When a specific egg version is requested and does not match what's
already in the cache, refetch it. Previously, the version in the cache
might have been used, despite the version difference. When the user
hasn't specified a version, the behaviour remains the same.

Fix the version comparison logic in `check-remote-version' so that an
egg is only fetched when a remote version is newer than the one in the
cache. Previously, this procedure was checking whether any remote
version is newer than the *requested* one (which might also be #f), but
it should actually indicate the opposite, i.e. it should return #t if
*no* remote version is newer than the local one (which should be the
cached version, not the requested version).

Signed-off-by: Peter Bex <peter@more-magic.net>