Restore read-string semantics when reading 0 bytes by not attempting to peek
We accidentally changed this behaviour when we changed read-string to
return #!eof on EOF.  The code was slightly restructured under the
assumption that calling peek-char will never hurt, and then return ""
when asking for zero bytes.

Logically, this is correct, but in some cases the FD might not be
ready.  In those cases you may run into a read timeout, but reading
zero bytes should arguably always succeed immediately, regardless of
port/FD state.

Signed-off-by: felix <felix@call-with-current-continuation.org>