Fix error in "chicken-status -cached" due to `map' with string argument
In 7a7d0be0 we added a clause to `list-egg-info' that maps over its
`dir' argument, but that may be reached when `dir' is a string (causing
an error). This procedure only actually needs to handle two cases,
installed egg-info files and egg or VERSION files in the cache, so we
simplify the procedure to fix this bug and remove some redundant logic.

Signed-off-by: Kooda <kooda@upyum.com>