Fix quotation in scripts generated by chicken-install and csc (#1515)
Instead of rolling our own quotation (which never works), we use the
tried and true "qs" procedure from (chicken process).  Unfortunately,
this procedure expects 'mingw32 as input, so we have to map platform
back to its original value again...

While testing this patch on UNIX with a path that contained spaces, I
also encountered a problem in the linking command which csc generates:
It was also misquoting its arguments.

While we're at it, we also remove the unused write-info procedure.  We
also ensure all "system" invocations have their quotes fixed on
Windows.

Many thanks to Kristian Lein-Mathisen for finding this bug.

Signed-off-by: felix <felix@call-with-current-continuation.org>

Added fix for double-quoting in "prepare-custom-command"