Allow multiple targets in "chicken-do" invocations
To handle selective deletion of files generated by egg-builds, we need a
way to make sure targets are re-built when one of several generated
files are missing or out of date. This patch generalizes chicken-do(1)
to allow multiple targets. The reordering of the arguments is inspired
by a proposal from Kooda.

Signed-off-by: Evan Hanson <evhan@foldling.org>