Extract most scrutinizer messages into separate functions
There are two main benefits from doing this:

 * Message formatting logic doesn't muddy the actual scrutinizer logic.
 * Giving names has a documentation benefit. If you refer to that name
   in tests you can grep for an examples of what might cause that error
   message.

Changes:

 * scrutinizer.scm (scrutinize): Copy report, report-notice to toplevel
   as report2, report-notice2
 * scrutinizer.scm: Add global *complain?*, needed by report2,
   report-notice2
 * scrutinizer.scm: Move multiples, node-source-prefix, location-name,
   fragment, pp-fragment under comment "Report helpers"
 * scrutinizer.scm (scrutinize): Remove report-error so 'errors'
   variable doesn't need to be made global

As a side effect (the symbol 1) now always gives a warning, which is for
the best; If you annotate a * return value to have some type with 'the,
there will be no warning without -verbose if the return value changes to
some explicit, and possibly incompatible, type.

A trivial example:

You want to write this:

  (+ 1 (foo))

You also know foo returns a fixnum, but foo does not have a type
annotation. So you might annotate the return value yourself:

  (+ 1 (the fixnum (foo)))

Now the + is likely specialized to fx+, which is fine.

Now the API for foo changes: foo returns a string and is given a type
annotation (-> string).

Now your code that annotates the return type is wrong. You are calling
fx+ with a string value. That will lead to undefined behaviour. Before
this commit you'll only see a warning when you use the -verbose flag.

Signed-off-by: Evan Hanson <evhan@foldling.org>