Print types in scrutinizer messages with pretty-print
Using `pp' to print types starts to shine when you're getting bigger
list, vector or procedure types.

Also, try to always print a relevant code fragment. This helps to find
the actual code the warning is originating from, or help debug the issue
when the actual code is just a macro call.

Signed-off-by: Evan Hanson <evhan@foldling.org>