Add unboxing pass to lfa2
After the lfa2 pass another pass is executed to eliminate unnecessary
boxing + unboxing of floating point intermediate values. The process
is roughly this: identify variables that are unassigned and are known
to contain flonums, count all accesses, then count all accesses of these
variables that are in direct operator position of an intrinsic that has
an unboxed variant and, if the number of accesses in unboxed position
is the same as the number of total accesses, then the variable
can be let-bound using a specialized construct (##core#let_float)
and all accesses be direct accesses (without any boxing/unboxing).
Results of unboxable intrinsics are boxed automatically
(using ##core#box_float), uses of ##core#inline_allocate on
unboxable intrinsics are converted to ##core#inline forms.

The lfa2 pass is now enabled at optimization levels 2 or higher.

Signed-off-by: Peter Bex <peter@more-magic.net>