Sanitize gensyms from scrutinizer test outputs
Instead of skipping tests that are sensitive to gensyms altogether, try
to sanitize the output.

For scrutinizer-message-format.scm, sanitize a and b because they are
used in typevars. These are removed once the scrutinizer sanitizes
typevars in messages internally again (fix for #1563 broke this).

Also, sanitize "scm:" so line number mismatches do not cause million
diff conflicts when adding/removing stuff to/from tests.