Fix empty line and memory leak in debugger GET_TRACE replies
Previously, all of the debugger's trace responses would include an empty
string in addition to the "real" entries, and the memory allocated for
the trace dump itself was never freed. The first issue is fixed by
sending only the newline-terminated chunks of the trace buffer (which
avoids including the empty string between the final newline and the
subsequent null terminator), and the second issue is fixed by simply
freeing the trace buffer at the end.

Signed-off-by: felix <felix@call-with-current-continuation.org>