Split "-no-trace" and "-debug-info" behaviours
This allows the "-no-trace" and "-debug-info" flags to be used together
with the intended effect, i.e. that debug events are still generated but
call traces are omitted. Previously, call traces would be generated at
"-d3" even when "-no-trace" was specified on the command line, since the
debugger's event handler would call C_trace() unconditionally whenever a
call event was encountered. Instead of leaving the C_trace() call there
we move it into the generated procedure body, and also make sure it
precedes the call to the C_debugger() so that the most recent item in
the call trace always matches the current procedure when the debugger is
triggered in response to a call event.

This also restores line number information to the traces, which were
unintentionally removed in ef5fbf34.

Signed-off-by: felix <felix@call-with-current-continuation.org>