Use consistent file extension for object files on Cygwin
On Cygwin, a different file extension was used for object files in csc
and chicken-install, resulting in missing object files and resultant
linking failures. In csc.scm "obj" was selected, but in egg-compile.scm,
and thus in chicken-install, we used "o" instead. This change makes them
match by using the same `software-version' check in both places.