When interning "qualified" symbols, convert them to regular symbols
This is a preparatory step in fixing #1077.  Once we have a
bootstrapping compiler that reads qualified symbols and interns them
as "##foo#bar" instead of as "\003foobar", it will also no longer emit
code with strings that are encoded in the old style.

The hack must be in the interning step so that the new runtime will
still be compatible with an old compiler: that compiler will still
generate code like "\003foobar".  Without this hack, we'd get errors
like "\003syscar is unbound" if we'd only change the reading of these
symbols.

Luckily, uninterned symbols don't matter for this.

This is a total hack which is also very wasteful and dumb because it
will malloc a new string whenever it encounters a qualified symbol.
But that's only in the intermediate bootstrapping compiler: after that,
when compiling a new CHICKEN, there should be no more dequalification
happening.

Signed-off-by: felix <felix@call-with-current-continuation.org>