Drop support for old-style qualified symbols (fixes #1077)
All checks for qualified symbols are either removed or replaced with
checks for keywords as those are the only remaining things that are
still encoded in the old style.

The procedure ##sys#symbol->qualified-string is no longer necessary,
instead we can simply call ##sys#symbol->string.

This also bumps BINARYVERSION to ensure new runtimes aren't mixed with
old code.

Signed-off-by: felix <felix@call-with-current-continuation.org>