* support.scm (constant-form-eval): Simplify logic
Change the code so 'k' is only called from tail position.

This simplifies the handling of case where the apply call causes an
exception. In the old code, this would cause a call to 'k' from a
non-tail position with ok value of #f. This would be handled in the
optimizer by returning the original n1. This is returned to
constant-form-eval as the value for the results variable. This causes
the first cond clause to fire (the one with the TODO comment), and 'k'
is called again.

Also, remove the form and msg arguments to 'k' as those are not used.

Signed-off-by: Peter Bex <peter@more-magic.net>