Add line number info for some forms introduced by reader
This adds entries to the line number database for the forms introduced
by the reader when it encounters "#$" (location), "#+" (cond-expand),
and the various quotation sigils. Previously, such forms would either
have no line number information (when occuring at the toplevel), or the
info would be inaccurate (when occuring inside another form, it would
use the line number of the outermost list).

Signed-off-by: Peter Bex <peter@more-magic.net>