Fix arguments to scrutiny reporting procedure for `append'
This commit fixes two issues introduced by 8301457, which updated the
scrutinizer's special case handling for `append' to use the new
`r-proc-call-argument-type-mismatch' procedure to report problems.

Firstly, the arguments for the call node and its type were flipped,
leading to an error when printing what was expected to be a node.

Secondly, the first item in the `arg-types' list was not resolved to a
type, so a `##core#the/result' node was passed instead. This wouldn't
cause an error, but it would give a confusing report.

Finally, for clarity, rename the `arg-types' variable to `args' within
that procedure (since it actually refers to nodes and not types), and
the `arg1' variable to `arg1-t' (which is a type and not a node).

Signed-off-by: Peter Bex <peter@more-magic.net>