Change representation of keywords
- Keywords are no longer encoded with a leading NUL byte.  This allows
   us to have proper read/write invariance of symbols and keywords and
   ensures we don't return #t for keyword? on symbols starting with \0.
- Keywords are now kept in symbol table completely separate from the
   one used for actual symbols.
- The plist of a keyword is now #f instead of '().  This is the one
   thing we can use to differentiate a keyword without attempting to
   look it up in the keyword table (which won't work while GC'ing).
- In order to be able to decide in which table to intern a symbol,
   when encoding literals, keywords and symbols are prefixed with a
   special byte: A \1-prefixed literal is decoded as a regular symbol,
   a \2-prefixed literal is read as a keyword.
- When persisting or unpersisting a symbol, loop through *all* the
   symbol tables when trying to locate the symbol.

Another small change is in how keywords are converted to nodes by the
compiler; originally they would (accidentally) be represented as
 ##core#variable nodes.  The intention was to auto-quote them in
canonicalize-expression, but due to how the cond was placed this
result would be thrown away and would be converted into ##core#variable
instead.

This is the first step towards fixing #1578.  For bootstrapping
reasons, this current implementation still accepts NUL-prefixed
symbols as keywords.  There is backwards compatible support code which
detects such symbols and interns them into (or looks them up in) the
keyword table instead.

Signed-off-by: felix <felix@call-with-current-continuation.org>