Fix types.db entries for posix file procedures and change file-truncate
We've made many of these procedures accept either strings (naming the
file), fixnums (indicating a file descriptor) or a port.  The port was
missing from several procedure entries in the types database.

The file-truncate procedure was an odd one out, it still only accepted
a file name or a descriptor; this has been fixed by also accepting a
port now.

This fixes #1609, by Robert Jensen

Signed-off-by: Evan Hanson <evhan@foldling.org>