Pass full macro name to `##sys#compiler-syntax-hook'
Previously, the unresolved macro name passed to the hook, meaning it
didn't include a module namespace and it was impossible to differentiate
two macros of the same name from different modules. This change just
passes the fully-qualified name of the syntax to the hook instead so
that reported compiler-syntax statistics ("-debug S") are more accurate.

Signed-off-by: Peter Bex <peter@more-magic.net>