Always regenerate import libraries when chicken-do is used
When building an egg, `chicken-do' considers the modification times of
import libraries when determining whether to build the files from which
they're generated generated. So, if `csc' skips regenerating an import
library because its imports haven't changed, the file that emits it will
be stuck in an outdated state and will always be rebuilt, unnecessarily.
This can lead to more extra work than is saved by not emitting the
import library in the first place, so introduce a flag to disable this
optimisation and use it when `chicken-do' is involved.

Signed-off-by: Kooda <kooda@upyum.com>