Fix how define-foreign-type defines its conversion procedures
It used to call "define", but this may not be available in the current
macro environment (or refer to something else than Scheme "define"),
so it was changed in de342aacd5bd18a6533e4eb4e04c2955e0f02959 to use
the ##core#set compiler form instead.  This would cause problems when
define-foreign-type was used inside a module, because later references
to the type would trigger an "undefined export" error, which would
cause the module to be unresolvable.

These procedures are gensymed, so they should not actually be
registered for export, but this restores the old behaviour based on
how "define" works at toplevel.  The compiler form already checks that
it is used at toplevel, so we may safely make the assumption that this
won't expand to an internal define in a lambda.

Signed-off-by: Evan Hanson <evhan@foldling.org>