Having keywords as a subtype of symbols is problematic.  This is most obvious with identifiers, which may be any symbol, but cannot be keywords.  Keywords also no longer have plists, so it makes less and less sense to treat these two object types as almost the same.
There are several Schemes which treat them as distinct types.  Having
keywords be a subtype of symbols is more of a old school Lisp thing,
like treating () and 'nil as the same, falsy values.  It's cleaner to
separate them, this also allows the scrutinizer to treat them more
cleanly as different things.

(Patch by Peter Bex, manually applied by Felix, due to git-fatigue)

Signed-off-by: felix <felix@call-with-current-continuation.org>