Fix obscure bug triggered by marking keywords persistable
When a keyword has already been read at runtime, and then compiled
code is invoked which calls C_h_intern_kw, the keyword will already
exist.  This is when it is marked as persistent because there's an
lf[] referring to it, so it can't be garbage collected.

The C_i_persist_symbol() call would then bail out because the offered
object is not a symbol but a keyword (the check was done with
C_i_check_symbol).

This caused several eggs depending on "foreigners" to break, because
it exports a macro in the import library which happens to use a
keyword.

Signed-off-by: Kooda <kooda@upyum.com>