Always pipe-quote symbols starting or ending with a colon
This is necessary, as shown by the failure of the srfi-128 egg: this
egg is compiled with keyword-style #:none in the compiler, which would
mean the types file emitted contains symbols like :type-test:, which
is read as a keyword when either other keyword style is set (and the
default is suffix).  All s-expressions written by one program should
be readable by another program, regardless of keyword style.

Signed-off-by: felix <felix@call-with-current-continuation.org>