Fix the most of #1604 by restoring rewrites dropped in  61af5f1
To make this work, we must remove the numeric operator rewrites
for (* *) in types.db, and instead rely on the rewrites in
c-platform.scm (which were already there), because specialization
happens before optimization.  So in the generic case, we'll end up
with the same code, but in fixnum arithmetic we'll end up with the
fixnum-specific versions.

Unfortunately, this still means that if the scrutinizer detects that
arguments are known to be integers (but not certain to be fixnums),
we'll generate calls to the more generic C_s_a_u_i_integer_... C
functions.  If they're known to be fixnums, we'll generate calls to
C_a_i_fixnum_..., which can overflow into bignums.  This is still not
optimal in fixnum arithmetic mode, because in that mode we'd want to
unsafely ignore overflow.

(Patch originally by Peter Bex)

Signed-off-by: felix <felix@call-with-current-continuation.org>