Emit C99 constants for +nan.0 and [+-]inf.0 `##core#float' nodes
The unboxing pass added in 79cf7427 introduces a `##core#float' type to
the closure-converted language, for which floating point literals are
emitted inline. This also needs to handle NaN and infinite values. For
these, we use the NAN and INFINITY constants defined by C99. We also
adds a cast to "double" for all inline flonums.

Fixes #1626.

Signed-off-by: Peter Bex <peter@more-magic.net>