Distinguish between IEEE fp positive and negative zero
To be able to read these numbers, we need a bit of a hacky workaround
because we read integers and then convert them to inexact when needed,
but of course proper integers don't distinguish between positive and
negative zero.

To write these numbers, we need to use signbit(f) instead of checking
whether the number is negative in a "normal" way, because -0.0 is not
smaller than 0.

To compare them, we have the bizarre rule that = will not distinguish,
while equal? and eqv? will.

Fixes #1627, thanks to John Cowan for pointing out this regression from
CHICKEN 4.x (CHICKEN 4 with the numbers egg has the same bug though).

Signed-off-by: Evan Hanson <evhan@foldling.org>