When contracting/inlining procedure calls, catch argument-list mismatch.
Previously a mismatch in the number of arguments to inlined procedures
would trigger an error. Recent changes in the compiler seem to have
exposed optimization oppurtunities that may result in more inlining
and show that the error may not be correct in all situations. This
patch simply aborts the inlining attempt when the arguments don't
match, with no error or warning shown (the available call site
information will be insufficient in nearly all cases).

Fixes #1630

Signed-off-by: Peter Bex <peter@more-magic.net>