Handle import libraries with static linkage egg property
For fully-static extensions, we need to both emit import libraries (so
they can be installed to support batch compilation) and compile module
registration code into the resulting objects (so the modules work when
imported in eval from statically-compiled programs).

This isn't possible with the current options, so add a flag and make the
`compile-module-registration` setting tri-valued: yes, no, and unset (to
keep the default behaviour). The egg toolchain then uses this flag to
compile static objects as usual while also emitting import libraries
when the static-linkage egg property is used.

Signed-off-by: felix <felix@call-with-current-continuation.org>