Replace car/cdr/null? on rest args with direct argvector references where possible
In the generated C code, we don't need C_build_rest() to dynamically
build a list containing everything in the argvector beyond the named
arguments.  Instead, we can try to detect references to positional
list accesses of the rest argument and convert those directly into
accesses of the argvector.  This means we build up less stack, causing
less GC pressure, which should reduce the number of minor GCs in tight
loops involving procedures with rest arguments.

The change introduces three new forms into the core language,
specifically for accessing rest arguments: ##core#rest-{car,cdr,null?}

These new forms represent a chain of calls to (cd...dr <rest-var>)
culminating in either a car, cdr or null? call.

When traversing a chain of rest-cdr calls, variables holding
intermediate cdrs may be eliminated, because only the final rest-cdr
or rest-car or rest-null? call matters.

When we see (if (null? <rest-cdr>) '() (rest-cdr <rest-cdr>)), the
variable which holds the result is marked as a rest-cdr variable.
This allows us to eventually eliminate any intermediate cdr calls on
the rest list.  This pattern is common in hand-rolled code, but it is
also generated by let-optionals*, which is in turn used by #!optional.
This catches the majority of rest arg usages.

In analyze-expression, the rest variable in ##core#rest-cdr nodes is
marked as captured to avoid total elimination of it.  This is
necessary, so that in closure conversion we still know that its home
closure is one that accepts rest arguments, and when inlining (see
below).

We also need to propagate rest-cdr to aliased variables, so that extra
"let"s don't block rest-cdr optimizations from happening.

One complicating factor is that the optimizer replaces calls to
scheme#{car,cdr,null?} with ##core#inline forms very early on.
This happens after the very first analysis.  So, we cannot match
directly on scheme#{car,cdr,null?}, but need to mark calls that look
like (##core#inline "C_i_{car,cdr,nullp}") instead.

Another tricky thing is that procedures with rest args may be inlined
or contracted.  When this happens, the "home" procedure of the rest arg
changes, so we can't replace references to rest arg cdrs with direct
argvector references anymore.  Therefore, we must rewrite the procedure
body when inlining it and potentially even re-introduce the rest
variable in a let binding.

To illustrate, let's look at an example with optional arguments and
how it will be optimized:

(lambda (#!optional (a 1) (b 2) (c 3))
  (print a b c))

this is equivalent to:

(lambda rest
  (let-optionals* rest ((a 1)
                        (b 2)
                        (c 3))
    (print a b c)))

and this is equivalent to and will get optimized as follows:

(lambda rest
  (let* ((a (if (null? rest) 1 (car rest)))
         (pre-b (if (null? rest) '() (cdr rest)))
	 (b (if (null? pre-b) 2 (car pre-b)))
	 (pre-c (if (null? pre-b) '() (cdr pre-b)))
	 (c (if (null? pre-c) 3 (car pre-c))))
    (print a b c)))

=={track rest-cdr call chain and replace with ##core#rest-... nodes}==>

(lambda rest
  (let* ((a (if (##core#rest-null? rest 0) 1 (##core#rest-car rest 0)))
         (pre-b (if (##core#rest-null? rest 1) '() (##core#rest-cdr rest 1)))
	 (b (if (##core#rest-null? rest 1) 2 (##core#rest-car rest 1)))
	 (pre-c (if (##core#rest-null? rest 2) '() (##core#rest-cdr rest 2)))
	 (c (if (##core#rest-null? rest 2) 3 (##core#rest-car rest 2))))
    (print a b c)))

=={eliminate unreferenced variables}==>

(lambda rest
  (let* ((a (if (##core#rest-null? rest 0) 1 (##core#rest-car rest 0)))
	 (b (if (##core#rest-null? rest 1) 2 (##core#rest-car rest 1)))
	 (c (if (##core#rest-null? rest 2) 3 (##core#rest-car rest 2))))
    (print a b c)))

Which, in C, basically translates to
(print <C_get_rest_arg(0)> <C_get_rest_arg(1)> <C_get_rest_arg(2)>)

This is how it's supposed to work conceptually.  The actual expansion
involves more LET variables, and the replacement happens in multiple
steps.

Also note that incorrect code like (null? (cdr (cdr (cdr rest)))) which
will normally crash with an error if rest does not contain at least 3
items will now simply return #t.  This is unfortunate but not a huge
deal considering this should be rare and also allowed by the spec I
think ("it is an error" doesn't mean "has to raise an exception").
Any other accesses of car or cdr beyond the list's end *are* translated
to code which will result in a runtime error, though.

This change should go a long way to improving #1623

Signed-off-by: felix <felix@call-with-current-continuation.org>