Disallow exporting variables defined with define-external
When finalizing modules, explicitly check in compiled code whether
the export refers to a variable defined with "define-external".
Since such variables represent foreign memory, they do not follow Scheme
semantics and any changes to them will not be reflected by an associated
exported variable. Note that this is different for functions, these
can not change spontaneously as compared to (say) a volatile external
variable.

Signed-off-by: Peter Bex <peter@more-magic.net>