Catch self-referencing variables in `letrec' form
"(letrec ((x x)) ...)" results in and endless recursion in the optimizer, since a letrec-bound
variable ends up being replacable by itself. The intermediate form generated by expanding letrec
in the compiler binds "x" first to (##core#undefined) which the compiler uses to assume
the variable has no value, leading to (in this context) somewhat questionable conclusion.
Since this is a rather tricky part of the analysis framework of the compiler, and due to the fact
that catching the situation earlier (during canonicalization) gives a more useful error message,
"letrec" expands now into a use of the internal form "##core#with-forbidden-refs", which marks
references to variables temporarily as invalid, until the code walk encounters
a lambda form.

"letrec" expands into a complex "let" form and is then walked again, during which the information
about the original letrec is lost, so we need an intermediate form that also takes care of
variable renaming. "##core#with-forbidden-refs" is dropped as soon as it is walked.

A test from syntax-tests.scm that actually used such a self-reference has been dropped.

No changes have been made to the interpreter. I think it's ok for the compiler to be more
picky and give better error detection.

Signed-off-by: felix <felix@call-with-current-continuation.org>
Signed-off-by: Peter Bex <peter@more-magic.net>