Check if there _is_ a value or syntax binding before warning and declaring it undefined
This unbreaks the build of CHICKEN itself, which "abuses" the
possibility of having both a procedure and a define-inline version of
the same procedure in tweaks.scm.

a) Check the list of unexportable stuff _after_ checking if there's a
   regular or syntax identifier defined for that export inside the module.

b) Add types to the list of things checked that cannot be exported.

Signed-off-by: megane <meganeka@gmail.com>