Print more information about why an identifier cannot be exported
After change:

Warning: Cannot export `a-type-alias' because it is a type abbreviation.

Warning: Cannot export `an-inline' because it is an inlined function.

Warning: Cannot export `a-constant' because it is a constant.

Warning: Cannot export `a-foreign' because it is a foreign variable.

Warning: Exported identifier `a-undefined' has not been defined.

When trying to compile this:
(module
 mod
 (a-type-alias an-inline a-constant a-foreign a-undefined)
 (import scheme)
 (cond-expand
  (chicken-5 (import (chicken base) (chicken type)
		     (chicken foreign)))
  (else (import chicken)))

 (define-type a-type-alias fixnum)
 (define-inline (an-inline) 1)
 (define-constant a-constant 2)
 (define-foreign-variable a-foreign int)
 )

Signed-off-by: Evan Hanson <evhan@foldling.org>