Don't infer types for foreign lambdas from foreign type with retconv/argconv
This is invalid, as shown by #1649.  Ideally we'd defer the type
inference of foreign types with retconv/argconv to the result of
scrutinizer's analysis of the retconv/argconv procedure's result, if
known.  To do so would be very tricky, as the way the types are
determined now is at macro-expansion time, which is before scrutiny.

So for now, we assign '* as the type if there's argconv or retconv.

Signed-off-by: Evan Hanson <evhan@foldling.org>