Improve heap shrinkage factor.
The old calculation was confused because it took a percentage
of a percentage.

This adds a new option to set heap shrinkage watermark and uses
that as-is to decide when to change the heap size.

By default, the heap will be shrunk by 50% when the used amount
is at 25% of the heap.  This looks closest to what seems to be
the intended behaviour of the code, and is also what was suggested
in #1379 by Sven Hartrumpf.

Signed-off-by: felix <felix@call-with-current-continuation.org>