Compile c-object components with proper flags
Egg components of type c-object should be compiled with the same C compiler flags
as C files generated by chicken, but currently the "-s" option for shared
objects is not given to c-objects when compiled in dynamic mode. This patch
adds the flag, fixing #1659.

This shortcoming was reported by Thomas Chust.

Signed-off-by: Kooda <kooda@upyum.com>