Fix remaining rest-argument bug #1658
After contracting/inlining a procedure, the database would still
contain information about rest-cdr variables, and the optimizations
for rest argument accessors would be incorrectly applied, because the
rest argument no longer is a rest argument but an explicitly consed
list in a local variable.

The rest cdr information on the variables is now cleared after
inlining.  To make this work, we keep track of all the derived rest
cdr notes in the database so that when a rest argument is moved
around, we can find all its aliases/derived cdrs.

Also add a new file containing test cases for this specific
optimization because it is trickier than it seems at first.

Signed-off-by: felix <felix@call-with-current-continuation.org>