Simplify really_mark to be a bit less branchy
Instead of checking the GC mode and doing subtly different but
functionally identical things in each mode, pass in the limits
and top pointer from the caller, reclaim().  There, the limits
are assigned when the gc mode is set.

NOTE: This does *NOT* make it faster.  It has no measurable impact at
all.  Hopefully, it should make it easier to understand what is going
on here.