Also convert remark() into really_mark calls.
This allows us to drop the really_remark function and related
multi-line macro and remark_system_globals(), which was the source of
a lot of code duplication with mark_system_globals().  This had the
risk of these two getting out of sync.

Again, no performance benefit to speak of, just simplification of the
code we have.