Fix a few platform-specific issues with current-process-millisecond
On Windows, clock() will return (clock_t)(-1) after about 24.8 days
(see https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/clock?view=vs-2019)
Instead, use GetTickCount64(), which returns the milliseconds since
system startup time as a 64-bit quantity, which does not have these
problems.  Unfortunately, the old mingw32 (which is close to
unmaintained) does not support this.  For compatibility, we will
use GetTickCount() there.

The C_NONUNIX macro is defined only on Mingw32 in chicken.h, so change
the #ifdef guard around this code to check on __MINGW32__, to make it
clear what it is checking for (for example, Haiku is not UNIX, but we
don't define that macro there).  If we need to support other non-UNIX
platforms, we should add specific guards per platform.

While we're at it, on Windows, also record the current value at
startup, so that we can tighten the definition of
current-process-milliseconds to state that it always returns the
number of milliseconds since process startup.  This means less
variance across platforms and makes the functionality less "vague".

Finally, split up the startup time into seconds and milliseconds, to
make the time since startup more precise; otherwise you get seemingly
random values when retrieving this due to only looking at the tv_sec
part of the gettimeofday return value.

Signed-off-by: felix <felix@call-with-current-continuation.org>